+++ isy-logging/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.0
 - `IFS-228`: Einführung von Spring Boot
+- `IFS-177`: Batch-Appender entfernt
 
 # 1.8.0
 - `IFS-225`: Prüfung der Logeinträge auf maximale Länge und ggfs. Kürzung dieser